refactor: Score, Result vo 패키지로 구분
