+++ src/main/java/domain/vo/Result.java
@@ -1,4 +1,6 @@
-package domain.model;
+package domain.vo;
+
+import java.util.Objects;
 
 public class Result {
 
@@ -22,6 +24,24 @@ public class Result {
 
     public void addDefeat() {
         defeat++;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        final Result result = (Result) o;
+        return getVictory() == result.getVictory() && getDraw() == result.getDraw()
+            && getDefeat() == result.getDefeat();
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(getVictory(), getDraw(), getDefeat());
     }
 
     public int getVictory() {
