+++ src/test/java/domain/model/ResultTest.java
@@ -2,6 +2,7 @@ package domain.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.vo.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
