+++ src/test/java/domain/model/ScoreTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import domain.type.Letter;
 import domain.type.Suit;
+import domain.vo.Score;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
