+++ src/test/java/domain/service/BlackJackResultMakerTest.java
@@ -6,7 +6,7 @@ import domain.model.Card;
 import domain.model.Cards;
 import domain.model.Dealer;
 import domain.model.Player;
-import domain.model.Result;
+import domain.vo.Result;
 import domain.type.Letter;
 import domain.type.Suit;
 import java.util.List;
